requirements.txt
action-msgs==1.2.2
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.9.0
ament-cmake-test==1.3.12
ament-copyright==0.12.14
ament-cppcheck==0.12.14
ament-cpplint==0.12.14
ament-flake8==0.12.14
ament-index-python==1.4.0
ament-lint==0.12.14
ament-lint-cmake==0.12.14
ament-package==0.14.1
ament-pep257==0.12.14
ament-uncrustify==0.12.14
ament-xmllint==0.12.14
angles==1.15.0
builtin-interfaces==1.2.2
composition-interfaces==1.2.2
cv-bridge==3.2.1
demo-nodes-py==0.20.5
diagnostic-msgs==4.9.0
domain-coordinator==0.10.0
example-interfaces==0.9.3
examples-rclpy-executors==0.15.4
examples-rclpy-minimal-action-client==0.15.4
examples-rclpy-minimal-action-server==0.15.4
examples-rclpy-minimal-client==0.15.4
examples-rclpy-minimal-publisher==0.15.4
examples-rclpy-minimal-service==0.15.4
examples-rclpy-minimal-subscriber==0.15.4
geometry-msgs==4.9.0
image-geometry==3.2.1
interactive-markers==2.3.2
laser-geometry==2.4.0
launch==1.0.10
launch-ros==0.19.12
launch-testing==1.0.10
launch-testing-ros==0.19.12
launch-xml==1.0.10
launch-yaml==1.0.10
lifecycle-msgs==1.2.2
logging-demo==0.20.5
map-msgs==2.1.0
message-filters==4.3.8
# Editable install with no version control (my-test-package==1.0)
-e /home/vsrivast/SPEAR/onboarding_proj_2_rc_car/build/my-test-package
nav-msgs==4.9.0
osrf-pycommon==2.1.6
packaging==25.0
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pygame==2.6.1
python-can==4.6.1
python-qt-binding==1.1.2
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.5
# Editable install with no version control (rc-car-controller==0.0.0)
-e /home/vsrivast/SPEAR/onboarding_proj_2_rc_car/build/rc_car_controller
rcl-interfaces==1.2.2
rclpy==3.3.17
rcutils==5.1.7
resource-retriever==3.1.3
rmw-dds-common==1.6.0
ros2action==0.18.14
ros2bag==0.15.15
ros2cli==0.18.14
ros2component==0.18.14
ros2doctor==0.18.14
ros2interface==0.18.14
ros2launch==0.19.12
ros2lifecycle==0.18.14
ros2multicast==0.18.14
ros2node==0.18.14
ros2param==0.18.14
ros2pkg==0.18.14
ros2run==0.18.14
ros2service==0.18.14
ros2topic==0.18.14
rosbag2-interfaces==0.15.15
rosbag2-py==0.15.15
rosgraph-msgs==1.2.2
rosidl-adapter==3.1.7
rosidl-cli==3.1.7
rosidl-cmake==3.1.7
rosidl-generator-c==3.1.7
rosidl-generator-cpp==3.1.7
rosidl-generator-py==0.14.5
rosidl-parser==3.1.7
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.7
rosidl-typesupport-introspection-cpp==3.1.7
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
sensor-msgs==4.9.0
sensor-msgs-py==4.9.0
shape-msgs==4.9.0
sros2==0.10.7
statistics-msgs==1.2.2
std-msgs==4.9.0
std-srvs==4.9.0
stereo-msgs==4.9.0
teleop-twist-keyboard==2.4.0
tf2-geometry-msgs==0.25.16
tf2-kdl==0.25.16
tf2-msgs==0.25.16
tf2-py==0.25.16
tf2-ros-py==0.25.16
tf2-sensor-msgs==0.25.16
tf2-tools==0.25.16
topic-monitor==0.20.5
trajectory-msgs==4.9.0
turtlesim==1.4.2
typing_extensions==4.15.0
unique-identifier-msgs==2.2.1
visualization-msgs==4.9.0
wrapt==1.17.3
